refactor(kissagotchi): separate logic into modules (state, minigame, ui, particles)
