+++ games_repo/games/kissagotchi/src/particles.rs
@@ -1,0 +1,20 @@
+use macroquad::prelude::*;
+
+pub struct Particle {
+    pub x: f32,
+    pub y: f32,
+    pub vx: f32,
+    pub vy: f32,
+    pub life: f32,
+    pub max_life: f32,
+    pub text: String,
+    pub color: Color,
+}
+
+impl Particle {
+    pub fn update(&mut self, dt: f32) {
+        self.x += self.vx * dt;
+        self.y += self.vy * dt;
+        self.life -= dt;
+    }
+}
