+++ games_repo/games/music_editor/Makefile
@@ -1,4 +1,4 @@
-VERSION=26.5.31.273
+VERSION=26.5.31.274
 
 
 .PHONY: build serve test lint
